+++ src/execution/execution_manager.cpp
@@ -164,7 +164,7 @@ void QlManager::select_from(std::unique_ptr<AbstractExecutor> executorTreeRoot, 
             } else if (col.type == TYPE_FLOAT) {
                 col_str = std::to_string(*(float *)rec_buf);
             }else if (col.type == TYPE_BIGINT) {
-                col_str = std::to_string(*(long long *)rec_buf);
+                col_str = std::to_string(*(int64_t *)rec_buf);
             } else if (col.type == TYPE_STRING) {
                 col_str = std::string((char *)rec_buf, col.len);
                 col_str.resize(strlen(col_str.c_str()));
