+++ src/execution/executor_abstract.h
@@ -68,7 +68,7 @@ public:
         else if (col.type == TYPE_FLOAT)
             ret.set_float(*(float *)dat);
         else if (col.type == TYPE_BIGINT)
-            ret.set_bigint(*(long long *)dat);
+            ret.set_bigint(*(int64_t *)dat);
         else if (col.type == TYPE_STRING)
         {
             std::string tmp(data, len);
